Drew a Rectangular Prism
I wrote a function to draw a prism based on specified points and the type of prism that it is. I did a line version and a solid version. The program draws the solid version first (in green) followed by the line version (in black). This can act as the body of our dino (and it can be adjusted if needed). Still need to add legs and a head and all that (as more rectangular prisms probably) and translate it around, but this should give us a good starting point (at least until we cover hidden object removal on Tuesday).